telegram-social-hub/backend/requirements.txt
aiogram>=3.3.0
python-dotenv
